Feature 2: refactor
