+++ internal/utils/split_slice.go
@@ -1,0 +1,26 @@
+package utils
+
+import "errors"
+
+// SplitSlice splits slice to several batches with size equals batchSize (except last)
+func SplitSlice(input []int, batchSize int) ([][]int, error) {
+	if input == nil {
+		return nil, nil
+	}
+
+	if batchSize < 1 {
+		return nil, errors.New("batchSize must be greater than 0")
+	}
+
+	var batched [][]int
+
+	for i := 0; i < len(input); i += batchSize {
+		rightBound := i + batchSize
+
+		if rightBound > len(input) {
+			rightBound = len(input)
+		}
+		batched = append(batched, input[i:rightBound])
+	}
+	return batched, nil
+}
